requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-settings==2.5.2

python-docx==1.1.2

# DB
SQLAlchemy==2.0.34
psycopg[binary]==3.2.2

# Qdrant + Elastic
qdrant-client==1.11.3
elasticsearch==8.15.1

# Embeddings (E5)
sentence-transformers==3.0.1
torch==2.4.0
python-multipart==0.0.9
transformers==4.44.2
